// src/app/models/cliente.interface.ts
export class ClienteUsuario{
    idusuario: number;
    ci: string;
    expedido: string;
    nombre: string;
    paterno: string;
    materno: string;
    celular: string;
    direccion: string;
    //producto = new Product;

}
/*export class Product{
    idproducto: number;
    nombre: string;
    descripcion: string;
    stock: number;
    minimo: number;
    maximo: number;
    vencimiento: number;
    precio: number;
    disponible: string;
    estado: string;
    peso: number;
    fecha: Date;

}*/